requirements.txt
backports.tarfile==1.2.0
gradio==4.44.0
httptools==0.6.1
ipykernel==6.29.5
jaraco.collections==5.1.0
langchain-chroma==0.1.4
langchain-experimental==0.3.0
langchain-openai==0.2.0
pandas-datareader==0.10.0
pip-chill==1.0.3
plotly==5.24.1
pygwalker==0.4.9.10
pypdf==5.0.0
sentence-transformers==3.1.1
streamlit==1.39.0
tomli==2.0.1
watchfiles==0.24.0
yahoo-fin==0.8.9.1
yfinance==0.2.54 #the latest version. Otherwise it would return emty data
pysqlite3-binary
